Implement JavaFX mediaplayer

+++ LLSIG/src/main/java/LLSIG/BeatTiming.java
@@ -1,10 +1,10 @@
 package main.java.LLSIG;
 
 public class BeatTiming {
-	long offset;
+	double offset;
 	int bpm;
 	boolean active;
-	BeatTiming(long offset,int bpm) {
+	BeatTiming(double offset,int bpm) {
 		this.offset=offset;
 		this.bpm=bpm;
 		this.active=true;
